requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
decouple==0.0.7
Django==4.2
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fastapi==0.95.1
filetype==1.2.0
flatbuffers==23.3.3
frozenlist==1.3.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.28.0
importlib-metadata==6.4.1
llvmlite==0.39.1
Markdown==3.4.3
mpmath==1.3.0
multidict==6.0.4

networkx==3.1
numba==0.56.4
numpy==1.23.5
onnxruntime==1.14.1
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pdf2image==1.16.3
Pillow==9.5.0
platformdirs==3.2.0
pooch==1.7.0
protobuf==4.22.3
pydantic==1.10.7
PyJWT==2.6.0
PyMatting==1.1.8
pyreadline3==3.4.1
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
rembg==2.0.33
requests==2.28.2
scikit-image==0.19.3
scipy==1.10.1
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.26.1
stripe==5.4.0
svgwrite==1.4.3
sympy==1.11.1
tifffile==2023.4.12
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
watchdog==3.0.0
yarl==1.9.1
zipp==3.15.0



gunicorn==20.0.4